+++ src/common/avc_es_parser.h
@@ -237,7 +237,6 @@ public:
   int64_rational_c const &get_par() const;
   std::pair<int64_t, int64_t> const get_display_dimensions(int width = -1, int height = -1) const;
 
-protected:
   bool parse_slice(memory_cptr const &nalu, slice_info_t &si);
   void handle_sps_nalu(memory_cptr const &nalu);
   void handle_pps_nalu(memory_cptr const &nalu);